fix(config): preserve serde default schema_version when both configs omit it
When neither user nor project config specifies schema_version, the merge
path previously forced it to 0 via unwrap_or(0). Now only inserts an
explicit schema_version when at least one source specifies it, letting
serde's default_schema_version() (= CURRENT_SCHEMA_VERSION) apply.

Also splits config_merge_tests.rs from config_tests.rs to stay under
the 800-line monolith guard limit.
